Move GenericCallbackFuture so it can be used in production code
It's a useful utility to have when dealing with callbacks.
