+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/proto/BookkeeperInternalCallbacks.java
@@ -23,12 +23,14 @@ package org.apache.bookkeeper.proto;
 
 import io.netty.buffer.ByteBuf;
 
+import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.RejectedExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 
+import org.apache.bookkeeper.client.BKException;
 import org.apache.bookkeeper.client.BookieInfoReader.BookieInfo;
 import org.apache.bookkeeper.client.LedgerEntry;
 import org.apache.bookkeeper.client.LedgerHandle;
@@ -136,6 +138,21 @@ public class BookkeeperInternalCallbacks {
                 statsLogger.registerFailedEvent(MathUtils.elapsedNanos(startTime), TimeUnit.NANOSECONDS);
             }
             cb.operationComplete(rc, result);
+        }
+    }
+
+    /**
+     * Generic callback future.
+     */
+    public static class GenericCallbackFuture<T>
+        extends CompletableFuture<T> implements GenericCallback<T> {
+        @Override
+        public void operationComplete(int rc, T value) {
+            if (rc != BKException.Code.OK) {
+                completeExceptionally(BKException.create(rc));
+            } else {
+                complete(value);
+            }
         }
     }
